refactor(headless): extract trajectory/tool-failure digest into prompt-candidate-digest.ts (#1409)
Move the digest pipeline out of prompt-candidate-loop.ts into a new sibling
prompt-candidate-digest.ts: the digest types (TrajectoryDigest,
TrajectoryToolCallDigest, TrajectoryToolFailureDigest,
ExtractTrajectoryDigestInput), the entry point extractTrajectoryDigest,
the cross-task aggregation renderToolFailureSummary, and the per-event
parser toolFailureDigest with its private helpers (extractToolFailureDigests,
passedResultsByTask, functionCallDigest/WithId, compareToolFailures,
argsPreview, promptSafeToken).

stripPromptOnlyToolFailures and renderToolFailureSummary are exported, since
the retained renderMetaAgentPrompt still uses them to feed the meta-agent
prompt. readRuntimeEventsJsonl is also exported, since it has real IO/parse
semantics and is shared by the retained scanRuntimeEventsForRewardHack;
prompt-candidate-loop.ts imports it back (single direction, no cycle).
modelVisibleStrings/stringValues stay in prompt-candidate-loop.ts, as they
are used only by scanRuntimeEventsForRewardHack.

prompt-candidate-loop.ts re-exports extractTrajectoryDigest and the four
digest types to preserve the public API, and imports the helpers it still
needs from the new module. The trivial pure guards isRecord/isNotFound stay
as private copies in each file, matching the existing codebase convention.
No behavior change; existing tests pass unchanged.
